Fixed declaration of borCityHash_64().

+++ boruvka/hfunc.h
@@ -84,7 +84,7 @@ uint32_t borCityHash_32(const void *buf, size_t size);
  * CityHash 64-bit hash function.
  * Taken from https://code.google.com/p/cityhash.
  */
-uint32_t borCityHash_64(const void *buf, size_t size);
+uint64_t borCityHash_64(const void *buf, size_t size);
 
 #ifdef __cplusplus
 }
